requirements.txt
pynmea2==1.15.0
pyserial==3.5
setproctitle==1.2.3
